Devendran m patch 5 (#10)
* doc(base): Add more instructions to README

This is a detailed section of the commit message where you can add more info about the change being committed. This is one of the recommended SemVer v2.0.0 and Conventitional Commit standards combination

This is an optional footer message where you can specify the PR

* feat: Add release-it process

To automate release-it

* Update package.json

Update package.json to include instructions

* Update package.json

Update package.json to include instructions. With updates

* Update package.json

* Update package.json to add autoGenerate

* Release 1.1.4